refactor: cube provider
