--- src/context/cube/index.ts
@@ -1,3 +1,0 @@
-export { default as CubeProvider } from "./CubeProvider";
-export * from "./CubeContext.types";
-export * from "./CubeContext.hooks";
