+++ src/context/hooks/index.ts
@@ -1,1 +1,2 @@
 export { default as useAuth } from "./useAuth";
+export { default as useCube } from "./useCube";
