+++ src/context/index.ts
@@ -1,1 +1,2 @@
 export { AuthProvider } from "./AuthContext";
+export { CubeProvider } from "./CubeContext";
