+++ src/layouts/BaseLayout/components/Drawer.tsx
@@ -24,7 +24,7 @@ import {
 import { NavLink, useLocation } from "react-router-dom";
 import { VERSION } from "src/consts";
 import { useAuth } from "src/context/hooks";
-import { useCube } from "src/context/cube";
+import { useCube } from "src/context/hooks";
 
 function getRoutes(isAdmin: boolean) {
   return [
