+++ src/pages/client/ImportPage/components/ImportedDataPage.tsx
@@ -15,7 +15,7 @@ import json from "src/assets/images/json.webp";
 import xls from "src/assets/images/xls.svg";
 import { ConfirmDialog } from "src/components";
 import { useAuth } from "src/context/hooks";
-import { useCube } from "src/context/cube";
+import { useCube } from "src/context/hooks";
 import { functions } from "src/firebase";
 import { getError } from "src/utils";
