+++ src/pages/client/Scorecard/Page.tsx
@@ -3,7 +3,7 @@ import { useCallback, useMemo, useState } from "react";
 import { Alert, Grid } from "@mui/material";
 import { ICubeData, IScorecardData } from "@shared/models";
 import { PageHeader } from "src/components";
-import { useCube } from "src/context/cube";
+import { useCube } from "src/context/hooks";
 import { useDocumentMetadata } from "src/hooks";
 import {
   updateStoringScorecardDataRow,
